fixed possible type change after API change in transformations

+++ pysigraph/scripts/sigraph/sulci_registration/independent_tag_with_registration.py
@@ -229,7 +229,7 @@ class Tagger(object):
                 X.append(Xs)
                 groups += [id] * len(Xs)
             id += 1
-        return numpy.vstack(X), numpy.array(groups)
+        return numpy.vstack(X).astype(float), numpy.array(groups)
 
 
 class GaussianTagger(Tagger):
